apps/backend_demo/requirements.txt
fastapi
uvicorn
python-multipart
typhoon-ocr
Pillow
pypdf
python-docx
google-genai
python-dotenv